+++ internal/clients/hcs-ama-api-spec/models/hashicorp_cloud_consulama_ama_billing_settings_prices_price_dimension_tier.go
@@ -6,12 +6,12 @@ package models
 // Editing this file might prove futile when you re-run the swagger generate command
 
 import (
-	strfmt "github.com/go-openapi/strfmt"
-
+	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
 )
 
 // HashicorpCloudConsulamaAmaBillingSettingsPricesPriceDimensionTier Tier represents one billing tier of a price dimension.
+//
 // swagger:model hashicorp.cloud.consulama.ama.BillingSettings.Prices.PriceDimension.Tier
 type HashicorpCloudConsulamaAmaBillingSettingsPricesPriceDimensionTier struct {