+++ internal/clients/hcs-ama-api-spec/models/hashicorp_cloud_consulama_ama_create_federation_token_response.go
@@ -6,12 +6,12 @@ package models
 // Editing this file might prove futile when you re-run the swagger generate command
 
 import (
-	strfmt "github.com/go-openapi/strfmt"
-
+	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
 )
 
 // HashicorpCloudConsulamaAmaCreateFederationTokenResponse CreateFederationTokenResponse contains the new Consul Federation Token.
+//
 // swagger:model hashicorp.cloud.consulama.ama.CreateFederationTokenResponse
 type HashicorpCloudConsulamaAmaCreateFederationTokenResponse struct {