+++ internal/clients/hcs-ama-api-spec/models/hashicorp_cloud_consulama_ama_create_token_response.go
@@ -6,13 +6,13 @@ package models
 // Editing this file might prove futile when you re-run the swagger generate command
 
 import (
-	strfmt "github.com/go-openapi/strfmt"
-
 	"github.com/go-openapi/errors"
+	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
 )
 
 // HashicorpCloudConsulamaAmaCreateTokenResponse CreateTokenResponse contains the new Consul Master Token.
+//
 // swagger:model hashicorp.cloud.consulama.ama.CreateTokenResponse
 type HashicorpCloudConsulamaAmaCreateTokenResponse struct {