+++ internal/clients/hcs-ama-api-spec/models/hashicorp_cloud_consulama_ama_get_operation_request.go
@@ -6,12 +6,12 @@ package models
 // Editing this file might prove futile when you re-run the swagger generate command
 
 import (
-	strfmt "github.com/go-openapi/strfmt"
-
+	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
 )
 
 // HashicorpCloudConsulamaAmaGetOperationRequest See ConsulAMAService.GetOperation
+//
 // swagger:model hashicorp.cloud.consulama.ama.GetOperationRequest
 type HashicorpCloudConsulamaAmaGetOperationRequest struct {