+++ internal/clients/hcs-ama-api-spec/models/hashicorp_cloud_consulama_ama_version.go
@@ -6,13 +6,13 @@ package models
 // Editing this file might prove futile when you re-run the swagger generate command
 
 import (
-	strfmt "github.com/go-openapi/strfmt"
-
 	"github.com/go-openapi/errors"
+	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
 )
 
 // HashicorpCloudConsulamaAmaVersion Version is a consul cluster version used in upgrade scenarios.
+//
 // swagger:model hashicorp.cloud.consulama.ama.Version
 type HashicorpCloudConsulamaAmaVersion struct {